+++ zcash_client_sqlite/src/wallet/init/migrations.rs
@@ -690,6 +690,9 @@ pub(super) fn verify_network_compatibility<P: consensus::Parameters>(
                     consensus::NetworkType::Main => "mainnet",
                     consensus::NetworkType::Test => "testnet",
                     consensus::NetworkType::Regtest => "regtest",
+                    consensus::NetworkType::Ycash => "ycash",
+                    consensus::NetworkType::YcashTest => "ycash testnet",
+                    consensus::NetworkType::Pirate => "pirate",
                 };
                 return Err(WalletMigrationError::CorruptedData(format!(
                     "Network type mismatch: account UFVK is for {} but attempting to initialize for {}.",